pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "transcribee"
version = "0.1.0"
description = "Local-first URL-to-transcript web app with pluggable transcription backends"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
  "authlib>=1.6.4,<2.0.0",
  "fastapi>=0.116.0,<1.0.0",
  "itsdangerous>=2.2.0,<3.0.0",
  "uvicorn[standard]>=0.35.0,<1.0.0",
  "jinja2>=3.1.6,<4.0.0",
  "python-multipart>=0.0.20,<1.0.0",
  "pydantic-settings>=2.10.1,<3.0.0",
  "pydantic>=2.11.9,<3.0.0",
  "sqlmodel>=0.0.24,<1.0.0",
  "yt-dlp>=2026.1.1",
]

[project.optional-dependencies]
faster-whisper = ["faster-whisper>=1.1.1"]
openai-whisper = ["openai-whisper>=20231117"]
whisper-cpp = ["pywhispercpp>=1.2.0"]
openai-api = ["openai>=1.0.0"]
deepgram = []
dev = [
  "httpx>=0.28.1,<1.0.0",
  "pytest>=8.4.2,<9.0.0",
  "pytest-asyncio>=1.1.0,<2.0.0",
  "ruff>=0.12.11,<1.0.0",
]

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py310"
